+++ R/document.R
@@ -34,6 +34,7 @@
 cnd_document <- function(
   package = get_package(),
   registry = package,
+  # TODO check if connection objects are appropriately handled
   file = file.path("R", paste0(package, "-cnd-conditions.R")),
   cleanup = TRUE
 ) {
@@ -91,10 +92,10 @@ cnd_document <- function(
     )
   )
 
-  temp <- file()
-  on.exit(if (isOpen(temp)) close(temp), add = TRUE)
-  cat(text, sep = "\n", file = temp)
-  res <- c(paste("#'", readLines(temp)), "NULL")
+  con_doc <- file()
+  on.exit(if (isOpen(con_doc)) close(con_doc), add = TRUE)
+  cat(text, sep = "\n", file = con_doc)
+  res <- c(paste("#'", readLines(con_doc)), "NULL")
   res <- trimws(res)
 
   if (is.null(file)) {
@@ -119,26 +120,29 @@ cnd_document <- function(
     cnd_files <- cnd_files[cnd_files != file]
     cnd_files <- filter2(
       cnd_files,
-      function(p) {
-        identical(
-          readLines(p, n = 1, warn = FALSE),
-          gen_text
-        )
-      }
+      \(p) identical(readLines(p, n = 1, warn = FALSE), gen_text)
     )
 
     if (length(cnd_files)) {
       cnd(cond_cnd_generated_cleanup(cnd_files))
-      try(file.remove(cnd_files), silent = TRUE)
+      tryCatch(unlink(cnd_files), error = \(e) NULL)
     }
   }
 
-  if (file.exists(file) && identical(res, readLines(file))) {
+  temp_gen <- tempfile()
+  on.exit(if (file.exists(temp_gen)) unlink(temp_gen), add = TRUE)
+  file.create(temp_gen)
+  Sys.chmod(temp_gen)
+  con_gen <- file(temp_gen, open = "wb", encoding = "UTF-8")
+  on.exit(if (isOpen(con_gen)) close(con_gen), add = TRUE)
+  cat(res, sep = "\n", file = con_gen)
+
+  if (file.exists(file) && md5sum(temp_gen) == md5sum(file)) {
     return(invisible(file))
   }
 
   cnd(cond_cnd_generated_write(file))
-  cat(res, sep = "\n", file = file)
+  file.copy(temp_gen, file, overwrite = TRUE)
   invisible(file)
 }
 
@@ -167,8 +171,8 @@ cnd_section <- function(fun) {
     ),
     pkgs = collapse(
       vapply(
-        unique(vapply(conds, cget, NA_character_, "package")),
-        function(p) fmt("[{pkg}-cnd-conditions]", pkg = p),
+        unique(vapply(conds, \(cond) cget(cond, "package"), NA_character_)),
+        \(p) fmt("[{pkg}-cnd-conditions]", pkg = p),
         NA_character_
       )
     ),